Time: 8 ms (8.70%), Space: 11.8 MB (53.03%) - LeetHub

+++ 0035-search-insert-position/0035-search-insert-position.cpp
@@ -4,7 +4,7 @@ public:
         int s = 0;
         int e = nums.size()-1;
         
-        long long int mid;
+        int mid;
         
         while(s <= e) {
             mid = s + (e-s)/2;
